backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.2
matplotlib==3.8.2
seaborn==0.13.2
numpy==1.26.4